refactor: movies queries functions

+++ backend/src/useCases/DeleteMovieUseCase.ts
@@ -1,3 +1,4 @@
+import { AppError } from "../errors/AppError";
 import { IMoviesRepository } from "../repositories/IMoviesRepository";
 
 type DeleteMovieUseCaseDTO = {
@@ -12,6 +13,11 @@ export class DeleteMovieUseCase {
   }
 
   async execute({ id }: DeleteMovieUseCaseDTO) {
+    const movie = this.moviesRepository.findById({ id })
+    if (!movie) {
+      throw new AppError('Filme não encontrado', 404)
+    }
+
     // TODO: check if there is a session with it
 
     await this.moviesRepository.delete({ id })
